fix: use substring instead of regex capture for departure ICAO. (#129)
The previous implementation used the following regex to match departure ICAO: `[A-Z]{4}`. The problem with it is that airports that have numbers in its ICAO codes were not parsed correctly, thus raising an error in runtime. Since the departure ICAO code has a fixed location in the string being parsed, it's safe to just use substr to parse it.